Rework migrate command.
