--- src/migrate.js
@@ -1,144 +1,0 @@
-const commandUtils = require( './command-utils' );
-const envUtils = require( './env-utils' );
-const fs = require( 'fs-extra' );
-const path = require( 'path' );
-const chalk = require( 'chalk' );
-const { execSync } = require( 'child_process' );
-const { exec } = require( 'child_process' );
-const environment = require( './environment' );
-
-const help = function() {
-    const help = `
-Usage:  10updocker migrate OLD_PATH [ENVIRONMENT]
-
-Migrate a WP Local Docker V1 environment into a new WP Local Docker V2 environment. Before running this command, create a new environment using the \`10updocker create\` command.
-
-OLD_PATH: Should be the path to the old WP Local Docker environment
-ENVIRONMENT: The environment to migrate the V1 data into.
-
-
-Example:
-  10updocker migrate ~/sites/mysite
-`;
-    console.log( help );
-    process.exit();
-};
-
-const validateOldEnv = async function( oldEnv ) {
-    if ( ! await fs.exists( path.join( oldEnv, 'docker-compose.yml' ) ) ) {
-        console.error( `${ chalk.bold.red( 'Error: ' ) }Could not find a docker-compose.yml file in the path specified for the old environment!` );
-        process.exit();
-    }
-
-    if ( ! await fs.pathExists( path.join( oldEnv, 'data', 'db' ) ) ) {
-        console.error( `${ chalk.bold.red( 'Error: ' ) }Could not find MySQL data in the path specified for the old environment!` );
-        process.exit();
-    }
-
-    return true;
-};
-
-const stopOldEnv = async function( oldEnv ) {
-    console.log( 'Ensuring old environment is not running...' );
-    try {
-        execSync( 'docker-compose down', { stdio: 'ignore', cwd: oldEnv } );
-    } catch( ex ) {}
-};
-
-// Kind of like the one for gateway, but not using docker compose and adapted for this purpose
-const waitForDB = function( containerName ) {
-    const readyMatch = 'ready for connections';
-    return new Promise( resolve => {
-        const interval = setInterval( () => {
-            console.log( 'Waiting for mysql...' );
-            exec( `docker logs ${containerName}`, ( error, stdout, stderr ) => {
-                if ( error ) {
-                    console.error( 'Error exporting database!' );
-                    process.exit();
-                }
-
-                if ( stderr.indexOf( readyMatch ) !== -1 ) {
-                    clearInterval( interval );
-                    resolve();
-                }
-            } );
-        }, 1000 );
-    } );
-};
-
-const exportOldDatabase = async function( oldEnv, exportDir ) {
-    const dataDir = path.join( oldEnv, 'data', 'db' );
-    const parts = path.parse( oldEnv );
-    const base = `mysql-${parts.name}`;
-
-    // Just in case this failed and are retrying
-    try {
-        execSync( `docker stop ${base}`, { stdio: 'ignore' } );
-        execSync( `docker rm ${base}`, { stdio: 'ignore' } );
-    } catch( ex ) {}
-
-    try {
-        execSync( `docker run -d --rm --name ${base} -v ${dataDir}:/var/lib/mysql -v ${exportDir}:/tmp/export mysql:5`, { stdio: 'inherit' } );
-        await waitForDB( base );
-        console.log( 'Exporting old database' );
-        execSync( `docker exec ${base} sh -c "/usr/bin/mysqldump -u root -ppassword wordpress > /tmp/export/database.sql"`, { stdio: 'inherit' } );
-        execSync( `docker stop ${base}` );
-    } catch ( ex ) {}
-};
-
-const importNewDatabase = async function( env ) {
-    await environment.start( env );
-    const envPath = await envUtils.getPathOrError( env );
-
-    try {
-        console.log( 'Importing DB to new Environment' );
-        execSync( 'docker-compose exec --user www-data phpfpm wp db import /var/www/html/import/database.sql', { stdio: 'inherit', cwd: envPath } );
-    } catch ( ex ) {}
-};
-
-const copySiteFiles = async function( oldEnv, newEnv ) {
-    const envPath = await envUtils.getPathOrError( newEnv );
-    const wpContent = path.join( envPath, 'wordpress', 'wp-content' );
-    const oldWpContent = path.join( oldEnv, 'wordpress', 'wp-content' );
-
-    // Clear out all the current environment content
-    // Only doing wp-content for now, since otherwise we would need to keep wp-config.php... But what about customizations?
-    console.log( `Removing current files from ${wpContent}` );
-    await fs.emptyDir( wpContent );
-
-    console.log( 'Copying files from the old wp-content folder' );
-    await fs.copy( oldWpContent, wpContent );
-};
-
-const command = async function() {
-    if ( commandUtils.subcommand() === 'help' || commandUtils.subcommand() === undefined ) {
-        help();
-    } else {
-        const old = path.resolve( commandUtils.getArg( 1 ) );
-        let env = commandUtils.getArg( 2 );
-
-        // So that we don't prompt at every step...
-        if ( env === false || undefined === env || env.trim().length === 0 ) {
-            env = await envUtils.promptEnv();
-        }
-
-        await validateOldEnv( old );
-        await stopOldEnv( old );
-
-        const envPath = await envUtils.getPathOrError( env );
-
-        // So that the DB is already in the folder mounted to docker
-        const exportDir = path.join( envPath, 'wordpress', 'import' );
-        await fs.ensureDir( exportDir );
-
-        await exportOldDatabase( old, exportDir );
-        await importNewDatabase( env );
-        await copySiteFiles( old, env );
-
-        console.log( `${ chalk.bold.green( 'Success!' ) } Your environment has been imported!` );
-        console.log( ' - wp-config.php has not been changed. Any custom configuration needs to be manually copied' );
-        console.log( ' - If you need to run a search/replace, run `10updocker wp search-replace <olddomain> <newdomain>`' );
-    }
-};
-
-module.exports = { command, help };
